+++ src/errors.js
@@ -43,6 +43,12 @@ class CircuitOpenError extends BareAgentError {
   }
 }
 
+class MaxRoundsError extends BareAgentError {
+  constructor(message, opts = {}) {
+    super(message || 'Loop exceeded maximum rounds', { code: 'MAX_ROUNDS', retryable: false, ...opts });
+  }
+}
+
 module.exports = {
   BareAgentError,
   ProviderError,
@@ -50,4 +56,5 @@ module.exports = {
   TimeoutError,
   ValidationError,
   CircuitOpenError,
+  MaxRoundsError,
 };